Prepare release

+++ vspace/src/main/java/edu/asu/diging/vspace/core/services/impl/SpaceManager.java
@@ -1,20 +1,25 @@
 package edu.asu.diging.vspace.core.services.impl;
 
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 
 import javax.transaction.Transactional;
 
 import org.apache.tika.Tika;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
+import edu.asu.diging.vspace.core.data.ExhibitionRepository;
 import edu.asu.diging.vspace.core.data.ImageRepository;
+import edu.asu.diging.vspace.core.data.SpaceLinkRepository;
 import edu.asu.diging.vspace.core.data.SpaceRepository;
 import edu.asu.diging.vspace.core.data.display.SpaceDisplayRepository;
+import edu.asu.diging.vspace.core.data.display.SpaceLinkDisplayRepository;
 import edu.asu.diging.vspace.core.exception.FileStorageException;
 import edu.asu.diging.vspace.core.exception.SpaceDoesNotExistException;
 import edu.asu.diging.vspace.core.factory.IImageFactory;
@@ -24,9 +29,12 @@ import edu.asu.diging.vspace.core.model.ISpace;
 import edu.asu.diging.vspace.core.model.IVSImage;
 import edu.asu.diging.vspace.core.model.display.ISpaceDisplay;
 import edu.asu.diging.vspace.core.model.display.impl.SpaceDisplay;
+import edu.asu.diging.vspace.core.model.impl.Exhibition;
 import edu.asu.diging.vspace.core.model.impl.Space;
+import edu.asu.diging.vspace.core.model.impl.SpaceLink;
 import edu.asu.diging.vspace.core.model.impl.SpaceStatus;
 import edu.asu.diging.vspace.core.model.impl.VSImage;
+import edu.asu.diging.vspace.core.services.IExhibitionManager;
 import edu.asu.diging.vspace.core.services.IImageService;
 import edu.asu.diging.vspace.core.services.ISpaceManager;
 import edu.asu.diging.vspace.core.services.impl.model.ImageData;
@@ -57,6 +65,19 @@ public class SpaceManager implements ISpaceManager {
     @Autowired
     private IImageService imageService;
 
+    @Autowired
+    private SpaceLinkRepository spaceLinkRepo;
+
+    @Autowired
+    private IExhibitionManager exhibitionManager;
+
+    @Autowired
+    private ExhibitionRepository exhibitRepo;
+
+    @Autowired
+    private SpaceLinkDisplayRepository spaceLinkDisplayRepo;
+
+    private final Logger logger = LoggerFactory.getLogger(getClass());
     /*
      * (non-Javadoc)
      * 
@@ -114,7 +135,7 @@ public class SpaceManager implements ISpaceManager {
         returnValue.setElement(space);
         return returnValue;
     }
-    
+
     /*
      * (non-Javadoc)
      * 
@@ -140,7 +161,7 @@ public class SpaceManager implements ISpaceManager {
             spaceDisplay.setHeight(image.getHeight());
             spaceDisplay.setWidth(image.getWidth());
         }
-        
+
         CreationReturnValue returnValue = new CreationReturnValue();
         returnValue.setErrorMsgs(new ArrayList<>());
 
@@ -154,7 +175,7 @@ public class SpaceManager implements ISpaceManager {
     @Override
     public ISpace getSpace(String id) {
         Optional<Space> space = spaceRepo.findById(id);
-        if (space.isPresent()) {
+        if (space != null && space.isPresent()) {
             return space.get();
         }
         return null;
@@ -180,28 +201,83 @@ public class SpaceManager implements ISpaceManager {
         return spaces;
     }
 
+
     @Override
     public List<ISpace> getSpacesWithStatus(SpaceStatus status) {
         List<ISpace> spaces = new ArrayList<>();
         spaceRepo.findAllBySpaceStatus(status).forEach(s -> spaces.add(s));
         return spaces;
     }
-    
+
     /**
      * Method to delete space based on id
      * 
-     * @param id
-     *            if id is null throws exception, else delete corresponding
-     *            space
-     * @throws SpaceDoesNotExistException 
+     * @param id if id is null throws exception, else delete corresponding space
+     * @throws SpaceDoesNotExistException
      */
     @Override
-    public void deleteSpaceById(String id) throws SpaceDoesNotExistException {
-        try {
+    public void deleteSpaceById(String id) {
+        if(id != null) {
+            List<SpaceLink> spaceLinks = spaceLinkRepo.getLinkedSpaces(id);
+            List<SpaceLink> fromSpaceLinks = spaceLinkRepo.getLinkedFromSpaces(id);
+            Exhibition exhibition = (Exhibition) exhibitionManager.getStartExhibition();
+            // When space has other links attached to it
+            // To delete links that access to the space getting deleted and replacing it as null
+            for(SpaceLink spaceLink : fromSpaceLinks) {
+                spaceLink.setTargetSpace(null);
+                spaceLinkRepo.save(spaceLink);
+            }
+            // To delete the links on the space getting deleted
+            for(SpaceLink spaceLink : spaceLinks) {
+                spaceLinkDisplayRepo.deleteBySpaceLinkId(spaceLink.getId());
+            }
+            spaceLinkRepo.deleteBySourceSpaceId(id);
+            // If the space is startSpace, we delete the space from the exhibition first.
+            if(exhibition != null && exhibition.getStartSpace() != null
+                    && exhibition.getStartSpace().getId().equalsIgnoreCase(id)) {
+                exhibition.setStartSpace(null);
+                exhibitRepo.save(exhibition);
+            }
+            // When space has no other links attached to it
+            spaceDisplayRepo.deleteBySpaceId(id);
             spaceRepo.deleteById(id);
-        } catch (IllegalArgumentException | EmptyResultDataAccessException exception) {
-            throw new SpaceDoesNotExistException(exception);
         }
+    }
 
+    @Override
+    public List<SpaceLink> getOutgoingLinks(String id) {
+
+        return spaceLinkRepo.getLinkedSpaces(id);
+    }
+
+    @Override
+    public List<SpaceLink> getIncomingLinks(String id) {
+
+        return spaceLinkRepo.getLinkedFromSpaces(id);
+    }
+
+    @Override
+    public List<ISpace> getSpacesWithImageId(String imageId) {
+        if(imageId == null) {
+            return null;
+        }
+        Optional<VSImage> vsImage = imageRepo.findById(imageId);
+        if(!vsImage.isPresent()) {
+            return null;
+        }
+        List<ISpace> spaces = new ArrayList<>();
+        spaceRepo.findAllByImageId(imageId).forEach(space -> spaces.add(space));
+        return spaces;
+    }
+
+
+    @Override
+    public Iterable<Space> addIncomingLinkInfoToSpaces(Iterable<Space> spaces) {
+        Iterator<Space> iterator = spaces.iterator();
+        while(iterator.hasNext()) {
+            Space space = iterator.next();
+            space.setIncomingLinks( (spaceLinkRepo.getLinkedFromSpaces(space.getId())).size() > 0 ? true : false );
+        }
+        return spaces;
     }
 }
